+++ inst/doc/gganimate.R
@@ -1,9 +1,13 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   gganimate = list(
-    nframes = 50
+    nframes = 50,
+    width = 1000,
+    height = 650,
+    units = "px",
+    res = 144
   ),
   out.width = '100%'
 )
@@ -80,14 +84,14 @@ anim +
   enter_fade() + enter_drift(x_mod = -1) + 
   exit_shrink() + exit_drift(x_mod = 5)
 
-## ---- eval=requireNamespace('av', quietly = TRUE)-----------------------------
-# Video output
-animate(
-  anim + enter_fade() + exit_fly(y_loc = 1),
-  renderer = av_renderer()
-)
+## ----eval=requireNamespace('av', quietly = TRUE)------------------------------
+#  # Video output
+#  animate(
+#    anim + enter_fade() + exit_fly(y_loc = 1),
+#    renderer = av_renderer()
+#  )
 
-## ---- out.width=NULL----------------------------------------------------------
+## ----out.width=NULL-----------------------------------------------------------
 # Different size and resolution
 animate(
   anim + ease_aes(x = 'bounce-out') + enter_fly(x_loc = -1) + exit_fade(),